+++ .github/scripts/utils.sh
@@ -374,11 +374,15 @@ generate_test_suites()
    mkdir suite_files
    SUITE_LIST=$(find test/e2e -type f \
      -name "*_test.go" \
-   ! -name "hazelcast_backup_slow_test.go" \
-   ! -name "hazelcast_wan_slow_test.go" \
-   ! -name "hazelcast_high_availability_test.go" \
-   ! -name "hazelcast_high_load_test.go" \
-   ! -name "hazelcast_soak_test.go" \
+   ! -name "platform_persistence_test.go" \
+   ! -name "platform_persistence_resilience_test.go" \
+   ! -name "platform_wan_test.go" \
+   ! -name "platform_resilience_test.go" \
+   ! -name "platform_rolling_upgrade_test.go" \
+   ! -name "platform_rollout_restart_test.go" \
+   ! -name "platform_tiered_storage_resilience_test.go" \
+   ! -name "platform_tiered_storage_test.go" \
+   ! -name "platform_soak_test.go" \
    ! -name "custom_resource_test.go" \
    ! -name "client_port_forward_test.go" \
    ! -name "e2e_suite_test.go" \
