+++ .github/workflows/generate-test-report.yaml
@@ -74,7 +74,7 @@ jobs:
 
       - name: Create Custom Defects Classification
         run: |-
-          echo "[{\"name\":\"Backup Slow Tests\",\"matchedStatuses\":[\"broken\",\"failed\"],\"traceRegex\":\".*hazelcast_backup_slow_test.*\"},{\"name\":\"Expose Externally Tests\",\"matchedStatuses\":[\"broken\",\"failed\"],\"traceRegex\":\".*hazelcast_expose_externally_test.*\"},{\"name\":\"Map Custom Class Tests\",\"matchedStatuses\":[\"broken\",\"failed\"],\"traceRegex\":\".*hazelcast_map_custom_class_test.*\"},{\"name\":\"Map Persistence Tests\",\"matchedStatuses\":[\"broken\",\"failed\"],\"traceRegex\":\".*hazelcast_map_persistence_test.*\"},{\"name\":\"Map Tests\",\"matchedStatuses\":[\"broken\",\"failed\"],\"traceRegex\":\".*hazelcast_map_test.*\"},{\"name\":\"Persistence Tests\",\"matchedStatuses\":[\"broken\",\"failed\"],\"traceRegex\":\".*hazelcast_persistence_test.*\"},{\"name\":\"Hazelcast Tests\",\"matchedStatuses\":[\"broken\",\"failed\"],\"traceRegex\":\".*hazelcast_test.*\"},{\"name\":\"Hazelcast Wan Tests\",\"matchedStatuses\":[\"broken\",\"failed\"],\"traceRegex\":\".*hazelcast_wan_test.*\"},{\"name\":\"Management Center Tests\",\"matchedStatuses\":[\"broken\",\"failed\"],\"traceRegex\":\".*management_center_test.*\"},{\"name\":\"Phone Home Tests\",\"matchedStatuses\":[\"broken\",\"failed\"],\"traceRegex\":\".*phone_home_test.*\"}]" > allure-results/${{ inputs.WORKFLOW_ID }}/categories.json
+          echo "[{\"name\":\"Platform Persistence Tests\",\"matchedStatuses\":[\"broken\",\"failed\"],\"traceRegex\":\".*platform_persistence_test.*\"},[{\"name\":\"Platform Persistence Resilience Tests\",\"matchedStatuses\":[\"broken\",\"failed\"],\"traceRegex\":\".*platform_persistence_resilience_test.*\"},{\"name\":\"Expose Externally Tests\",\"matchedStatuses\":[\"broken\",\"failed\"],\"traceRegex\":\".*expose_externally_test.*\"},{\"name\":\"Map Tests\",\"matchedStatuses\":[\"broken\",\"failed\"],\"traceRegex\":\".*map_test.*\"},{\"name\":\"Platform Persistence Tests\",\"matchedStatuses\":[\"broken\",\"failed\"],\"traceRegex\":\".*platfrom_persistence_test.*\"},{\"name\":\"Hazelcast Tests\",\"matchedStatuses\":[\"broken\",\"failed\"],\"traceRegex\":\".*hazelcast_test.*\"},{\"name\":\"Wan Tests\",\"matchedStatuses\":[\"broken\",\"failed\"],\"traceRegex\":\".*wan_test.*\"},{\"name\":\"Management Center Tests\",\"matchedStatuses\":[\"broken\",\"failed\"],\"traceRegex\":\".*management_center_test.*\"},{\"name\":\"Phone Home Tests\",\"matchedStatuses\":[\"broken\",\"failed\"],\"traceRegex\":\".*phone_home_test.*\"},{\"name\":\"Backup Restore Tests\",\"matchedStatuses\":[\"broken\",\"failed\"],\"traceRegex\":\".*backup_restore_test.*\"},{\"name\":\"Cache Tests\",\"matchedStatuses\":[\"broken\",\"failed\"],\"traceRegex\":\".*cache_test.*\"},{\"name\":\"Jet Job Tests\",\"matchedStatuses\":[\"broken\",\"failed\"],\"traceRegex\":\".*jetjob_test.*\"},{\"name\":\"Jet Job Snapshot Tests\",\"matchedStatuses\":[\"broken\",\"failed\"],\"traceRegex\":\".*jetjobsnapshot_test.*\"},{\"name\":\"Multi Namespace Tests\",\"matchedStatuses\":[\"broken\",\"failed\"],\"traceRegex\":\".*multi_ns_test.*\"},{\"name\":\"Multimap Tests\",\"matchedStatuses\":[\"broken\",\"failed\"],\"traceRegex\":\".*multimap_test.*\"},{\"name\":\"Options Tests\",\"matchedStatuses\":[\"broken\",\"failed\"],\"traceRegex\":\".*options_test.*\"},{\"name\":\"Platform WAN Tests\",\"matchedStatuses\":[\"broken\",\"failed\"],\"traceRegex\":\".*paltform_wan_test.*\"},{\"name\":\"Platform Rolling Upgrade Tests\",\"matchedStatuses\":[\"broken\",\"failed\"],\"traceRegex\":\".*platform_rolling_upgrade_test.*\"},{\"name\":\"Platform Rollout Restart Tests\",\"matchedStatuses\":[\"broken\",\"failed\"],\"traceRegex\":\".*platform_rollout_restart_test.*\"},{\"name\":\"Platform Soak Tests\",\"matchedStatuses\":[\"broken\",\"failed\"],\"traceRegex\":\".*platform_soak_test.*\"},{\"name\":\"Platform Tiered Storage Resilience Tests\",\"matchedStatuses\":[\"broken\",\"failed\"],\"traceRegex\":\".*platform_tiered_storage_resilience_test.*\"},{\"name\":\"Queue Tests\",\"matchedStatuses\":[\"broken\",\"failed\"],\"traceRegex\":\".*queue_test.*\"},{\"name\":\"Replicated Map Tests\",\"matchedStatuses\":[\"broken\",\"failed\"],\"traceRegex\":\".*replicatedmap_test.*\"},{\"name\":\"Resilience Tests\",\"matchedStatuses\":[\"broken\",\"failed\"],\"traceRegex\":\".*resilience_test.*\"},{\"name\":\"Tiered Storage Tests\",\"matchedStatuses\":[\"broken\",\"failed\"],\"traceRegex\":\".*tiered_storage_test.*\"},{\"name\":\"Topic Tests\",\"matchedStatuses\":[\"broken\",\"failed\"],\"traceRegex\":\".*topic_test.*\"},{\"name\":\"User Code Tests\",\"matchedStatuses\":[\"broken\",\"failed\"],\"traceRegex\":\".*user_code_test.*\"}]" > allure-results/${{ inputs.WORKFLOW_ID }}/categories.json
 
       - name: Generate Test Report
         id: generate-report
